refactor: extract SRAM/RTC/State save functions into minarch_saves.c
Moves formatSavePath, SRAM_getPath/read/write, RTC_getPath/read/write,
state_slot, State_getPath/read/write, State_autosave, and State_resume
out of minarch.c into a dedicated minarch_saves.c module.

State_getPath is exposed publicly since Menu_updateState calls it directly.
